adding intitial models of the main structure
These models were sourced from https://github.com/yucnet/gym-dobot
The license for these are currently unknow and will need to be reviewed.

They seem to be the best models of the outer shell that I can find
related to a "dobot" like robot. They should make good models to 3d
print the outer structure with.